chore: sort scripts
