+++ src/scripts/indexBlogsQdrant.mjs
@@ -1,9 +1,10 @@
-// Major ref: https://js.langchain.com/docs/modules/indexes/vector_stores/integrations/pinecone
-import { PineconeClient } from "@pinecone-database/pinecone";
+// Major ref: https://js.langchain.com/docs/modules/indexes/vector_stores/integrations/qdrant
+
 import dotenv from "dotenv";
 import { Document } from "langchain/document";
 import { OpenAIEmbeddings } from "langchain/embeddings/openai";
-import { PineconeStore } from "langchain/vectorstores/pinecone";
+import {QdrantClient} from '@qdrant/js-client-rest';
+import { QdrantVectorStore } from "langchain/vectorstores/qdrant";
 import fs from "fs";
 import path from "path";
 
@@ -34,7 +35,7 @@ fileNames.forEach((fileName) => {
   const fileContent = fs.readFileSync(filePath, "utf8");
 
   const chunks = chunkString(fileContent, MAX_TOKENS);
-  
+
   chunks.forEach((chunk, index) => {
     lanchainDocs.push(new Document({
       metadata: { fileName, chunkIndex: index },
@@ -43,17 +44,14 @@ fileNames.forEach((fileName) => {
   });
 });
 
-const client = new PineconeClient();
-await client.init({
-  apiKey: process.env.PINECONE_API_KEY,
-  environment: process.env.PINECONE_ENVIRONMENT,
-});
-const pineconeIndex = client.Index(process.env.PINECONE_INDEX);
 
-await PineconeStore.fromDocuments(
-  lanchainDocs,
+const qdrantClient = new QdrantClient({ url: process.env.QDRANT_URL, apiKey: process.env?.QDRANT_API_KEY });
+
+await QdrantVectorStore.fromDocuments(
+  langchainDocs,
   new OpenAIEmbeddings({ openAIApiKey: process.env.OPENAI_API_KEY }),
   {
-    pineconeIndex,
+    client: qdrantClient,
+    collectionName: process.env.QDRANT_COLLECTION_NAME,
   }
 );
